Update project files to include new files added and fix some windows build issues.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@1311058 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/decaf/util/concurrent/locks/AbstractQueuedSynchronizerTest.cpp
@@ -669,7 +669,7 @@ void AbstractQueuedSynchronizerTest::testAwaitNanosTimeout() {
     AbstractQueuedSynchronizer::ConditionObject* c = mutex.newCondition();
     try {
         mutex.acquire(1);
-        long t = c->awaitNanos(100);
+        long long t = c->awaitNanos(100);
         CPPUNIT_ASSERT(t <= 0);
         mutex.release(1);
     } catch(Exception& ex) {
